refactor(db-builder): decouple InventoryManager from java-instrumentation-watcher internals (#454)
Move JavaagentInventoryManager to watcher-common so explorer-db-builder
can import it without depending on java-instrumentation-watcher internals.

- Add JavaagentInventoryManager to watcher-common/inventory_manager.py
- Slim java-instrumentation-watcher InventoryManager to extend it (backward compat)
- Update explorer-db-builder to import from watcher_common.inventory_manager
- Remove java-instrumentation-watcher from explorer-db-builder dependencies
