+++ ecosystem-automation/explorer-db-builder/pyproject.toml
@@ -7,13 +7,11 @@ dependencies = [
     "PyYAML>=6.0.1",
     "semantic-version>=2.10.0",
     "collector-watcher",
-    "java-instrumentation-watcher",
     "watcher-common",
 ]
 
 [tool.uv.sources]
 collector-watcher = { workspace = true }
-java-instrumentation-watcher = { workspace = true }
 watcher-common = { workspace = true }
 
 [project.scripts]
