+++ ecosystem-automation/watcher-common/pyproject.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 description = "Shared utilities for OpenTelemetry ecosystem watchers"
 requires-python = ">=3.11"
 dependencies = [
+    "PyYAML>=6.0.1",
     "semantic-version>=2.10.0",
 ]
